requirements.txt
django==4.2.18
djangorestframework==3.15.2
